+++ tests/assets/compilers/typescript/input.ts
@@ -1,0 +1,13 @@
+type FullName = string;
+
+interface User {
+    firstName: string;
+    lastName: string;
+}
+
+
+function getName(u: User): FullName {
+    return `${u.firstName} ${u.lastName}`;
+}
+
+let userName: FullName = getName({firstName: "Django", lastName: "Pipeline"});